Sequlize actualizado/script.js actualizado/express-validator integrado/Views actualizadas

+++ src/public/jscript/script.js
@@ -5,8 +5,12 @@ const fComment = document.getElementById('form-comment');
 const like = document.getElementById('like');
 const cLikes = document.getElementById('c-likes');
 const bDelete = document.querySelector('.btn-delete');
+const fBody = document.querySelector('.form-body');
+const error = document.querySelectorAll('.error-msg-i');
 
-iComment.addEventListener('click', createBox);
+if (iComment) {
+	iComment.addEventListener('click', createBox);
+};
 
 function createBox() {
 	const btnBox = document.createElement('DIV');
@@ -33,28 +37,42 @@ function createBox() {
 	iComment.removeEventListener('click', createBox);
 };
 
-like.addEventListener('click', (e)=>{
-	const imgId = like.getAttribute('data-id');
-	fetch(`/gallery/${imgId}/like`,{
-		method: 'POST',
-		body: JSON.stringify({}),
-		headers: {'Content-type': 'application/json'}
-	})
-		.then(res => res.json())
-		.then(res2 => cLikes.textContent = `${res2.likes}`)
-});
-
-bDelete.addEventListener('click', e=>{
-	e.preventDefault();
-	const res = confirm('Are you sure delete this image?');
-	if (res) {
-		const imgId = like.getAttribute('data-Id');
-		fetch(`/gallery/${imgId}`,{
-			method: 'DELETE',
-			headers: {'Content-type': 'application/json'},
-			redirect:'follow'
+if (like) {
+	like.addEventListener('click', (e)=>{
+		const imgId = like.getAttribute('data-id');
+		fetch(`/gallery/${imgId}/like`,{
+			method: 'POST',
+			body: JSON.stringify({}),
+			headers: {'Content-type': 'application/json'}
 		})
 			.then(res => res.json())
-			.then(res => window.location.href = res);
+			.then(res2 => cLikes.textContent = `${res2.likes}`)
+	});
+};
+
+if (bDelete) {
+	bDelete.addEventListener('click', e=>{
+		e.preventDefault();
+		const res = confirm('Are you sure delete this image?');
+		if (res) {
+			const imgId = bDelete.getAttribute('data-id');
+			fetch(`/gallery/${imgId}`,{
+				method: 'DELETE',
+				headers: {'Content-type': 'application/json'},
+				redirect:'follow'
+			})
+				.then(res => res.json())
+				.then(res => window.location.href = res);
+		};
+	});
+};
+
+if (error) {
+	for (const e of error) {
+		e.addEventListener('click', ()=>{
+			const id = e.getAttribute('data-id');
+			const element = document.getElementById(`${id}`);
+			fBody.removeChild(element);
+		});
 	};
-});
+};
